+++ src/scripts/content/scraper/scrapPage.js
@@ -4,7 +4,7 @@ import {extractPage} from "../extractor/extractPage";
 import {defineExportMethod} from "../export/defineExportMethod";
 import {updateClickIconCount} from "../../background/icon/clickCount/clickIconCountContext";
 import {safeExecute} from "../../shared/utils/jsShorteners";
-import {EXPORTER_FALLBACK_ACTION, EXTRACTOR_FALLBACK_ACTION} from "./fallbackActions";
+import {EXPORTER_FALLBACK_ACTION, EXTRACTOR_FALLBACK_ACTION} from "../utils/fallbackActions";
 import {formatFilename} from "../../shared/formatter/formatText";
 import {applyExtractorRules} from "../extractor/rules/applyRules";
 import converter from "../../shared/formatter/formatMarkdown";
