+++ src/scripts/content/utils/fallbackActions.js
